requirements.txt
retina-face
tensorflow
pillow
requests
scikit-image
pyyaml
tqdm
